create cosntants for dataSkills
