+++ vite-project/src/Pages/SkillSection.tsx
@@ -6,29 +6,9 @@ import {
   Button,
   Typography,
 } from "@mui/material";
-import boxOffice from "../assets/img/box office.webp";
-import filmDatabase from "../assets/img/film database.webp";
-import logicGame from "../assets/img/logic game.webp";
-import marketPlace from "../assets/img/marketplace.webp";
-import upgradePorfolio from "../assets/img/upgradePorfolio.jpeg";
-import menu from "../assets/img/menu.png";
-import notes from "../assets/img/notes.webp";
-import portfolio from "../assets/img/portfolio.webp";
-import {
-  SiHtml5,
-  SiCss3,
-  SiSass,
-  SiTailwindcss,
-  SiTypescript,
-  SiJavascript,
-  SiPhp,
-  SiReact,
-  SiFirebase,
-  SiMui,
-} from "react-icons/si";
+import DataSkills from "../constants/dataSkills";
 import { CheckedStateProps } from "../types";
 import usePageStyles from "../styles/stylePages";
-
 import { useIntl } from "react-intl";
 
 const SkillSection = ({
@@ -37,127 +17,6 @@ const SkillSection = ({
   checkedState: CheckedStateProps;
 }) => {
   const intl = useIntl();
-  const DataSkills = [
-    {
-      id: 1,
-      img: portfolio,
-      icon: [
-        <SiHtml5 style={{ fontSize: "20px", color: "#e44d26" }} />,
-        <SiSass style={{ fontSize: "20px", color: "#cd6799" }} />,
-        <SiJavascript style={{ fontSize: "20px", color: "#f7e025" }} />,
-      ],
-      technologies: ["html", "sass", "javascript"],
-      nameProject: `${intl.formatMessage({ id: "project.namePortfolio" })}`,
-      detailProject: `${intl.formatMessage({
-        id: "project.namePortfolioText",
-      })}`,
-      openProject: "https://github.com/katzohub/Portfolio-App",
-    },
-    {
-      id: 2,
-      img: menu,
-      icon: [
-        <SiHtml5 style={{ fontSize: "20px", color: "#e44d26" }} />,
-        <SiSass style={{ fontSize: "20px", color: "#cd6799" }} />,
-        <SiJavascript style={{ fontSize: "20px", color: "#f7e025" }} />,
-      ],
-      technologies: ["html", "sass", "javascript"],
-      nameProject: `${intl.formatMessage({ id: "project.restaurant" })}`,
-      detailProject: `${intl.formatMessage({
-        id: "project.restaurantText",
-      })}`,
-      openProject: "https://github.com/katzohub/Info-Food-App",
-    },
-    {
-      id: 3,
-      img: logicGame,
-      icon: [
-        <SiHtml5 style={{ fontSize: "20px", color: "#e44d26" }} />,
-        <SiCss3 style={{ fontSize: "20px", color: "#306af1" }} />,
-        <SiJavascript style={{ fontSize: "20px", color: "#f7e025" }} />,
-      ],
-      technologies: ["html", "css", "javascript"],
-      nameProject: `${intl.formatMessage({ id: "project.nameGame" })}`,
-      detailProject: `${intl.formatMessage({
-        id: "project.nameGameText",
-      })}`,
-      openProject: "https://github.com/katzohub/Game-Seven-App",
-    },
-    {
-      id: 4,
-      img: boxOffice,
-      icon: [
-        <SiPhp style={{ fontSize: "20px", color: "#7b7fb5" }} />,
-        <SiCss3 style={{ fontSize: "20px", color: "#306af1" }} />,
-      ],
-      technologies: ["php", "css"],
-      nameProject: `${intl.formatMessage({ id: "project.nameOffice" })}`,
-      detailProject: `${intl.formatMessage({
-        id: "project.nameOfficeText",
-      })}`,
-      openProject: "https://github.com/katzohub/Box-office-php",
-    },
-    {
-      id: 5,
-      img: filmDatabase,
-      icon: [
-        <SiReact style={{ fontSize: "20px", color: "#61dafb" }} />,
-        <SiJavascript style={{ fontSize: "20px", color: "#f7e025" }} />,
-        <SiCss3 style={{ fontSize: "20px", color: "#306af1" }} />,
-      ],
-      technologies: ["react", "css", "javascript"],
-      nameProject: `${intl.formatMessage({ id: "project.nameDatabase" })}`,
-      detailProject: `${intl.formatMessage({
-        id: "project.nameDatabaseText",
-      })}`,
-      openProject: "https://github.com/katzohub/Film-Database-App",
-    },
-    {
-      id: 6,
-      img: notes,
-      icon: [
-        <SiReact style={{ fontSize: "20px", color: "#61dafb" }} />,
-        <SiTypescript style={{ fontSize: "20px", color: "#007acc" }} />,
-        <SiTailwindcss style={{ fontSize: "20px", color: "#38bdf8" }} />,
-      ],
-      technologies: ["react", "tailwind", "typescript"],
-      nameProject: `${intl.formatMessage({ id: "project.nameNotes" })}`,
-      detailProject: `${intl.formatMessage({
-        id: "project.nameNotesText",
-      })}`,
-      openProject: "https://github.com/katzohub/REACT-TS",
-    },
-    {
-      id: 7,
-      icon: [
-        <SiReact style={{ fontSize: "20px", color: "#61dafb" }} />,
-        <SiTypescript style={{ fontSize: "20px", color: "#007acc" }} />,
-        <SiFirebase style={{ fontSize: "20px", color: "#ffcc32" }} />,
-      ],
-      technologies: ["react", "firebase", "javascript"],
-      img: marketPlace,
-      nameProject: `${intl.formatMessage({ id: "project.nameMarkPlace" })}`,
-      detailProject: `${intl.formatMessage({
-        id: "project.nameMarkPlaceText",
-      })}`,
-      openProject: "https://github.com/katzohub/marketplace-Mobile-App",
-    },
-    {
-      id: 8,
-      icon: [
-        <SiReact style={{ fontSize: "20px", color: "#61dafb" }} />,
-        <SiTypescript style={{ fontSize: "20px", color: "#007acc" }} />,
-        <SiMui style={{ fontSize: "20px", color: "#007fff" }} />,
-      ],
-      technologies: ["react", "typescript", "mui"],
-      img: upgradePorfolio,
-      nameProject: `${intl.formatMessage({ id: "project.portolioUpgrade" })}`,
-      detailProject: `${intl.formatMessage({
-        id: "project.portolioUpgradeText",
-      })}`,
-      openProject: "https://github.com/katzohub/Portfolio",
-    },
-  ];
 
   const isMobile = useMediaQuery("(max-width:800px)");
   const filteredProjects = checkedState.all
@@ -209,9 +68,11 @@ const SkillSection = ({
                   />
                   <div className={classes.skillCardHeading}>
                     <h2 className={classes.skillCardText}>
-                      {project.nameProject}
+                      {intl.formatMessage({ id: project.nameProject })}
                     </h2>
-                    <p style={{ fontSize: "15px" }}>{project.detailProject}</p>
+                    <p style={{ fontSize: "15px" }}>
+                      {intl.formatMessage({ id: project.detailProject })}
+                    </p>
 
                     <Box className={classes.skillCardContainer}>
                       {project.icon &&
